+++ Src/Common/Models/Earn/GetStakeRedeemOrderHistoryResult.cs
@@ -1,30 +1,52 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Newtonsoft.Json;
 
 namespace bybit.net.api.Models.Earn
 {
     public class GetStakeRedeemOrderHistoryResult
     {
-        public List<EarnOrderHistoryItem>? list { get; set; }
+        [JsonProperty("list")]
+        public List<EarnOrderHistoryItem>? List { get; set; }
+
+        [JsonProperty("nextPageCursor")]
+        public string? NextPageCursor { get; set; }
     }
 
     public class EarnOrderHistoryItem
     {
-        public string? coin { get; set; }
-        public string? orderValue { get; set; }
-        public string? orderType { get; set; }       // Redeem, Stake
-        public string? orderId { get; set; }
-        public string? orderLinkId { get; set; }
-        public string? status { get; set; }          // Success, Fail, Pending
-        public string? createdAt { get; set; }       // ms
-        public string? productId { get; set; }
-        public string? updatedAt { get; set; }       // ms
-        public string? swapOrderValue { get; set; }  // LST Onchain only
-        public string? estimateRedeemTime { get; set; } // Onchain only, ms
-        public string? estimateStakeTime { get; set; }  // Onchain only, ms
-    }
+        [JsonProperty("coin")]
+        public string? Coin { get; set; }
 
+        [JsonProperty("orderValue")]
+        public string? OrderValue { get; set; }
+
+        [JsonProperty("orderType")]
+        public string? OrderType { get; set; }
+
+        [JsonProperty("orderId")]
+        public string? OrderId { get; set; }
+
+        [JsonProperty("orderLinkId")]
+        public string? OrderLinkId { get; set; }
+
+        [JsonProperty("status")]
+        public string? Status { get; set; }
+
+        [JsonProperty("createdAt")]
+        public string? CreatedAt { get; set; }
+
+        [JsonProperty("productId")]
+        public string? ProductId { get; set; }
+
+        [JsonProperty("updatedAt")]
+        public string? UpdatedAt { get; set; }
+
+        [JsonProperty("swapOrderValue")]
+        public string? SwapOrderValue { get; set; }
+
+        [JsonProperty("estimateRedeemTime")]
+        public string? EstimateRedeemTime { get; set; }
+
+        [JsonProperty("estimateStakeTime")]
+        public string? EstimateStakeTime { get; set; }
+    }
 }